requirements.txt
# --- Core Data & API ---
pandas==2.2.3
requests==2.32.3
numpy==2.0.2

# --- HTML & Grafici ---
Jinja2==3.1.6
plotly==5.24.1
kaleido==0.1.0

# --- Automazione e Sicurezza ---
schedule==1.2.2
python-dotenv==1.0.1
